+++ docs/RefMan/conf.py
@@ -36,7 +36,7 @@ author = 'The Cryptol Team'
 # The short X.Y version
 version = ''
 # The full version, including alpha/beta/rc tags
-release = '3.4.1'
+release = '3.4.1.99'
 
 
 # -- General configuration ---------------------------------------------------